Project/requirements.txt
Django~=4.1.5
djongo~=1.3.6
pymongo~=4.3.3
dnspython~=2.3.0
djangorestframework~=3.14.0
bcrypt~=4.0.1
stripe~=5.0.0
xhtml2pdf~=0.2.9
python-dateutil~=2.8.2
pycountry~=22.3.5
whitenoise~=6.4.0